% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_environment_vars}
\alias{check_environment_vars}
\title{Check environment vars
Check if the required variables are set}
\usage{
check_environment_vars(vars)
}
\arguments{
\item{vars}{a list of environment variables}
}
\value{
NULL if okay
}
\description{
Check environment vars
Check if the required variables are set
}
