% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_table_helpers.R
\name{get_colour}
\alias{get_colour}
\title{Get a color based on a numeric value}
\usage{
get_colour(
  value,
  limit_red = LIMIT_RED,
  limit_orange = LIMIT_ORANGE,
  colours = COLOURS
)
}
\arguments{
\item{value}{Numeric value to evaluate}

\item{limit_red}{Numeric value that represents lower bound for colour red.}

\item{limit_orange}{Numeric value that represents lower bound for colour orange.}

\item{colours}{Named list that includes hex colour code values for red, green and orange.}
}
\value{
A color string or NULL if value is not numeric or NULL
}
\description{
Returns a color hex code depending on the value.
Used for conditional cell coloring in tables.
}
