% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_db_connection}
\alias{get_db_connection}
\title{get connection to database}
\usage{
get_db_connection(
  server_address = DB_SERVER,
  database = DB_DATABASE,
  username = DB_USERNAME,
  password = DB_PASSWORD,
  db_driver = DB_DRIVER,
  sqlite_path = SQLITE_DB_PATH
)
}
\arguments{
\item{server_address}{server address}

\item{database}{database name}

\item{username}{username to be used to login to the database}

\item{password}{password to be used to login to the database}

\item{db_driver}{sql db driver to be used to setup connection to database}

\item{sqlite_path}{sqlite path when using a local db}
}
\value{
a connection object to the specified database
}
\description{
get connection to database
}
