% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_js_click}
\alias{get_js_click}
\title{javascript function for getting clicked cell in a table}
\usage{
get_js_click(id, clickable_columns = NULL)
}
\arguments{
\item{id}{cell}
}
\value{
value of clicked cell
}
\description{
javascript function for getting clicked cell in a table
}
