% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table}
\alias{get_table}
\title{Get a reference to a table in the configured schema}
\usage{
get_table(con, table, schema = DB_SCHEMA)
}
\arguments{
\item{con}{Database connection}

\item{table}{Table name}

\item{schema}{Schema name (default: DB_SCHEMA). Ignored for SQLite.}
}
\value{
dplyr table reference
}
\description{
This function returns a dplyr reference to a table.
When using a SQLite database, the schema is ignored.
}
