% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input_selectors_helpers.R
\name{get_unique_column_values_in_table}
\alias{get_unique_column_values_in_table}
\title{Get unique column values table
Retrieves the unique column values for every column specified in a table.}
\usage{
get_unique_column_values_in_table(schema, table, columns)
}
\arguments{
\item{schema}{database schema}

\item{table}{database table}

\item{columns}{vector of column names}
}
\value{
Returns a list of unique values in database table for
specified columns
}
\description{
Get unique column values table
Retrieves the unique column values for every column specified in a table.
}
\examples{
get_unique_column_values_in_table(schema = "dbo",
                                            table = "table",
                                            columns = c("period", "group"))
}
