% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_table_helpers.R
\name{merge_coldefs}
\alias{merge_coldefs}
\title{Merge reactable::colDef() objects into one}
\usage{
merge_coldefs(...)
}
\arguments{
\item{...}{One or more `reactable::colDef()` objects. These are passed as arguments
to this function. Each `colDef()` object must contain uniquely named parameters.
Duplicate parameter names across the colDef arguments will cause an error.}
}
\value{
A single merged `reactable::colDef()` object.
}
\description{
Merges multiple `reactable::colDef()` objects into a single `reactable::colDef()` object.
}
\examples{
merge_coldefs(
  reactable::colDef(name = "Color", align = "center"),
  reactable::colDef(style = function(value) paste0("color:", value))
)
}
